% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_question_numbers.R
\name{add_question_numbers}
\alias{add_question_numbers}
\title{Add question numbers to ODK XLS Form}
\usage{
add_question_numbers(ODKFile, language = "English",
  maintype = "numbers", subtype = "numbers")
}
\arguments{
\item{ODKFile}{ODK form read in using read.odk}

\item{language}{language of label column to add numbers into}

\item{maintype}{how to label sections. Defaults to "numbers". Also can have
"LETTERS" or "letters"}

\item{subtype}{how to label questions within sections. Defaults to
"numbers". Also can have "LETTERS" or "letters"}
}
\description{
Take existing labels from an ODK XLS form and attach automatically generated
question numbers based on groups and questions within groups New numbers will
be generated for each new section (group or repeat) - nesting of sections not
currently implemented Questions will be assigned to the following ODK
elements: "integer","decimal", "select_one",
"select_multiple","text","date","datetime","time", "image","geopoint"
}
\examples{
#Form<-read.odk("data/EvaluationQuestionnaire2018.xlsx")
#Form \%>\% add_question_numbers()
}
\keyword{Number}
\keyword{ODK}
\keyword{Question}
\keyword{XLSForm}
