% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_reference_lists.R
\name{cross_reference_lists}
\alias{cross_reference_lists}
\title{cross reference the choices tab with information from the survey tab}
\usage{
cross_reference_lists(ODKFile, by = "label", language = "English")
}
\arguments{
\item{ODKFile}{ODK form read in using read.odk}

\item{by}{which column from survey tab to use as cross reference}

\item{language}{language of column (can be "" if no languages referred to in from)}
}
\description{
This is useful for checking and translation purposes to have the question
text next to the answer text to provide additional context and make sure
wording is appropriate. Multiple questions mapping to same list set will be concatenated
}
\examples{
#Form<-read.odk("data/EvaluationQuestionnaire2018.xlsx")
#Form \%>\% cross_reference_lists()
}
\keyword{ODK}
\keyword{lists}
\keyword{unsused}
