% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readodk.R
\name{read.odk}
\alias{read.odk}
\title{Read in an ODK XLS form into R}
\usage{
read.odk(ODKFile)
}
\arguments{
\item{ODKFile}{file location where XLSForm is saved}
}
\description{
Converts ODK form into list data frame of survey, choices and settings, for use in this package
}
\examples{
#Form<-read.odk("data/EvaluationQuestionnaire2018.xlsx")
#write.odk(Form,"data/EvaluationQuestionnaire2018_modified.xlsx")
}
\keyword{Number}
\keyword{ODK}
\keyword{Question}
\keyword{XLSForm}
