% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dropped_lists.R
\name{remove_dropped_lists}
\alias{remove_dropped_lists}
\title{Remove any lists from the choices tabs that do not appear in the main survey}
\usage{
remove_dropped_lists(ODKFile, verbose = TRUE)
}
\arguments{
\item{ODKFile}{ODK form read in using read.odk}

\item{verbose}{return message indicating whether to show message or just
return object}
}
\description{
These will be moved into a new tab called unused_choices. Standard ODK
validations check if all lists from survey are in choices but not the other
way around. The reverse is useful to prevent translation of obsolete lists
and for form checking
}
\examples{
#Form<-read.odk("data/EvaluationQuestionnaire2018.xlsx")
#Form \%>\% remove_dropped_lists()
}
\keyword{ODK}
\keyword{lists}
\keyword{unsused}
