% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.odkxls}
\alias{summary.odkxls}
\title{summary function for class type odkxls}
\usage{
\method{summary}{odkxls}(object)
}
\arguments{
\item{object}{imported xls form}
}
\description{
summary function for class type odkxls
}
\examples{
#Form<-read.odk("data/EvaluationQuestionnaire2018.xlsx")
#summary(Form)
}
\keyword{Number}
\keyword{ODK}
\keyword{Question}
\keyword{XLSForm}
