% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_blanks.R
\name{trigger_blanks}
\alias{trigger_blanks}
\title{Add acknowledgement fields at end of each question group to indicate if any of the questions have been left blank}
\usage{
trigger_blanks(ODKFile, groups = "all", language = "English",
  hinttext = "At least one question within this section has been left blank. Select OK if this has been skipped intentionally; otherwise please check responses")
}
\arguments{
\item{ODKFile}{ODK form read in using read.odk}

\item{groups}{add triggers to "all" groups (default) or to user specified list of groups}

\item{language}{language of hint column to add trigger hint}

\item{hinttext}{text that will appear in trigger messages}
}
\description{
This is useful if you do not want to force all questions to be required, but still want to make sure that all questions are being addressed and not left blank accidentally
}
\examples{
#Form<-read.odk("data/EvaluationQuestionnaire2018.xlsx")
#Form \%>\% trigger_blanks()
}
\keyword{Number}
\keyword{ODK}
\keyword{Question}
\keyword{XLSForm}
