% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeodk.R
\name{write.odk}
\alias{write.odk}
\title{Write an ODK object back to XLS form}
\usage{
write.odk(obj, file)
}
\arguments{
\item{obj}{R object}

\item{file}{location to write new XLS form}
}
\description{
Outputs form modified in R back to an XLS form
}
\examples{
#Form<-read.odk("data/EvaluationQuestionnaire2018.xlsx")
#write.odk(Form,"data/EvaluationQuestionnaire2018.xlsx")
}
\keyword{ODK}
\keyword{form}
\keyword{write}
